ニコニコのサムネのことは考えずにとりあえずフィットするように
