+++ src/util/canvas.ts
@@ -1,0 +1,8 @@
+export const createCanvas = (width: number, height: number) => {
+    const canvas = document.createElement('canvas')
+    canvas.width = width
+    canvas.height = height
+    const context = canvas.getContext('2d')!
+
+    return { canvas, context }
+};
